Explicit reference in Step 1 of Setup process
In Azure Cloud Shell, we have 2 options: Bash and PowerShell. If the user is using PowerShell by default, then Step 2 of Setup process will generate "The '<' operator is reserved for future use" error message. To avoid this, it's better to specify explicitly in Step 1 that the one has to switch to Bash.